pyproject.toml
[project]
name = "rag-self"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "chromadb>=1.5.9",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "modelscope>=1.36.3",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pypdf2>=3.0.1",
    "pytest>=9.0.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "sentence-transformers>=5.4.1",
    "sqlalchemy>=2.0.49",
    "sse-starlette>=3.4.1",
    "uvicorn[standard]>=0.46.0",
]
